atualizando configs e mensagens de log.

+++ graduacao/src/main/java/br/usp/exemplo/graduacao/servicos/AlunoServico.java
@@ -39,6 +39,9 @@ public class AlunoServico {
   }
 
   public void deletar(Long id) {
+    if(!alunoRepositorio.existsById(id)){
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Aluno não encontrado com o ID: " + id);
+    }
     alunoRepositorio.deleteById(id);
   }
